[otlib,usb] Move `UsbHub` from tests to otlib
Generic handling of USB hub is useful not only in tests but also
for opentitanlib itself.

(cherry picked from commit 53c3f6f47791c8a40c1658dce9bf43d6729a0005)
